backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
python-slugify==8.0.1
